WifiManager: reorg and split MQTT vs Wifi code
