+++ RFLink/RFLink.ino
@@ -25,9 +25,7 @@
 #include "5_Plugin.h"
 #include "6_WiFi_MQTT.h"
 #include "8_OLED.h"
-#if defined(RFLINK_OTA_ENABLED)
-  #include "ArduinoOTA.h"
-#endif // RFLINK_OTA_ENABLED
+#include "10_Wifi.h"
 
 #if (defined(__AVR_ATmega328P__) || defined(__AVR_ATmega2560__))
 #include <avr/power.h>
@@ -78,6 +76,7 @@ void setup()
   Serial.println(); // ESP "Garbage" message
   Serial.print(F("Arduino IDE Version :\t"));
   Serial.println(ARDUINO);
+#endif
 #ifdef ESP8266
   Serial.print(F("ESP CoreVersion :\t"));
   Serial.println(ESP.getCoreVersion());
@@ -87,33 +86,18 @@ void setup()
   Serial.println(F("Compiled on :\t\t" __DATE__ " at " __TIME__));
 
 
-#ifdef RFLINK_WIFIMANAGER_ENABLED
+#if defined(RFLINK_WIFIMANAGER_ENABLED) || defined(RFLINK_WIFI_ENABLED)
 RFLink::Wifi::setup();
 #endif // RFLINK_WIFIMANAGER_ENABLED
 
 #ifdef MQTT_ENABLED
-  #ifndef RFLINK_WIFIMANAGER_ENABLED
-    setup_WIFI();
-    start_WIFI();
-    #ifdef RFLINK_OTA_ENABLED
-      #ifdef RFLINK_OTA_PASSWORD
-        ArduinoOTA.setPassword(RFLINK_OTA_PASSWORD);
-      #endif
-      ArduinoOTA.begin();
-    #endif // RFLINK_OTA_ENABLED
-  #endif // RFLINK_WIFIMANAGER_ENABLED
-  setup_MQTT();
-  reconnect(1);
-#else
-  #ifndef RFLINK_WIFIMANAGER_ENABLED
-  setup_WIFI_OFF();
-  #endif // RFLINK_WIFIMANAGER_ENABLED
+  RFLink::Mqtt::setup_MQTT();
+  RFLink::Mqtt::reconnect(1);
 #endif // MQTT_ENABLED
-#endif // ESP32 || ESP8266
 
-  PluginInit();
-  PluginTXInit();
-  set_Radio_mode(Radio_OFF);
+PluginInit();
+PluginTXInit();
+set_Radio_mode(Radio_OFF);
 
 #if ((defined(ESP8266) || defined(ESP32)) && !defined(RFM69_ENABLED))
   show_Radio_Pin();
@@ -134,7 +118,7 @@ RFLink::Wifi::setup();
   splash_OLED();
 #endif
 #ifdef MQTT_ENABLED
-  publishMsg();
+  RFLink::Mqtt::publishMsg();
 #endif
   pbuffer[0] = 0;
   set_Radio_mode(Radio_RX);
@@ -143,11 +127,11 @@ RFLink::Wifi::setup();
 void loop()
 {
 #ifdef MQTT_ENABLED
-  checkMQTTloop();
+  RFLink::Mqtt::checkMQTTloop();
   sendMsg();
 #endif
 
-#ifdef RFLINK_WIFIMANAGER_ENABLED
+#if defined(RFLINK_WIFIMANAGER_ENABLED) || defined(RFLINK_WIFI_ENABLED)
 RFLink::Wifi::mainLoop();
 #endif
 
@@ -155,9 +139,6 @@ RFLink::Wifi::mainLoop();
   ArduinoOTA.handle();
 #endif
 
-#ifdef SERIAL2NET_ENABLED
-RFLink::Serial2Net::serverLoop();
-#endif // SERIAL2NET_ENABLED
 
 
 #ifdef SERIAL_ENABLED
@@ -178,7 +159,7 @@ void sendMsg()
     Serial.print(pbuffer);
 #endif
 #ifdef MQTT_ENABLED
-    publishMsg();
+    RFLink::Mqtt::publishMsg();
 #endif
 #ifdef OLED_ENABLED
     print_OLED();
